process_Tfs_accepted.js has just one function process_Tfs_accepted()
process_Tfs_accepted.js has just one function process_Tfs_accepted() This function will check that:
phase is 1 which means that the task in the taskflow-suggest phase where the task-helper is just coming up with a list of tasks and 2 means that the task-helper has actually started working on the task s4AprvTfs is not 1 means that because s4AprvTfs used to represent "submit taskflow-suggest for an approval" and when a taskflow-suggest is submitted for an approval s4AprvTfs gets set to a 1 so we are interested only in tasks where taskflow-suggest have not yet been submitted for an approval.
tfsStatusID is 5 means that because taskStatusID is set to a value of 5 when the task-helper is initially assigned to a task and they have not yet accepted the task. tfsStatusID can have a value of 5, 1, or 2. tfsStatusID is 5 when task-helper is just assigned task and has not yet accepted task
tfsStatusID is 1 when task-helper has accepted the task
tfsStatusID is 2 when task-helper has accepted the task and the 72 hour time limit to complete and submit their taskflow-suggest and ran out.